Break out FileReader to File.js
