+++ src/File.js
@@ -1,0 +1,54 @@
+const {EventTarget} = require('./Event');
+
+class FileReader extends EventTarget {
+  constructor() {
+    super();
+
+    this.result = null;
+  }
+
+  get onload() {
+    return _elementGetter(this, 'load');
+  }
+  set onload(onload) {
+    _elementSetter(this, 'load', onload);
+  }
+
+  get onerror() {
+    return _elementGetter(this, 'error');
+  }
+  set onerror(onerror) {
+    _elementSetter(this, 'error', onerror);
+  }
+
+  readAsArrayBuffer(file) {
+    this.result = file.buffer.buffer.slice(file.buffer.byteOffset, file.buffer.byteOffset + file.buffer.byteLength);
+
+    process.nextTick(() => {
+      this.dispatchEvent(new Event('load', {
+        target: this,
+      }));
+    });
+  }
+
+  readAsDataURL(file) {
+    this.result = 'data:' + file.type + ';base64,' + file.buffer.toString('base64');
+
+    process.nextTick(() => {
+      this.dispatchEvent(new Event('load', {
+        target: this,
+      }));
+    });
+  }
+
+  readAsText(file) {
+    this.result = file.buffer.toString('utf8');
+
+    process.nextTick(() => {
+      this.dispatchEvent(new Event('load', {
+        target: this,
+      }));
+    });
+  }
+}
+module.exports.FileReader = FileReader;
