+++ src/core.js
@@ -14,6 +14,8 @@ const {performance} = require('perf_hooks');
 
 const mkdirp = require('mkdirp');
 
+const {FileReader} = require('./File.js');
+
 const {XMLHttpRequest: XMLHttpRequestBase, FormData} = require('window-xhr');
 
 const fetch = require('window-fetch');
@@ -464,58 +466,6 @@ class DataTransferItem {
     const {data} = this;
     setImmediate(() => {
       callback(data);
-    });
-  }
-}
-
-class FileReader extends EventTarget {
-  constructor() {
-    super();
-
-    this.result = null;
-  }
-
-  get onload() {
-    return _elementGetter(this, 'load');
-  }
-  set onload(onload) {
-    _elementSetter(this, 'load', onload);
-  }
-
-  get onerror() {
-    return _elementGetter(this, 'error');
-  }
-  set onerror(onerror) {
-    _elementSetter(this, 'error', onerror);
-  }
-
-  readAsArrayBuffer(file) {
-    this.result = file.buffer.buffer.slice(file.buffer.byteOffset, file.buffer.byteOffset + file.buffer.byteLength);
-
-    process.nextTick(() => {
-      this.dispatchEvent(new Event('load', {
-        target: this,
-      }));
-    });
-  }
-
-  readAsDataURL(file) {
-    this.result = 'data:' + file.type + ';base64,' + file.buffer.toString('base64');
-
-    process.nextTick(() => {
-      this.dispatchEvent(new Event('load', {
-        target: this,
-      }));
-    });
-  }
-
-  readAsText(file) {
-    this.result = file.buffer.toString('utf8');
-
-    process.nextTick(() => {
-      this.dispatchEvent(new Event('load', {
-        target: this,
-      }));
     });
   }
 }
